Paginando outras entidades

+++ ProjectEntra21/src/Domain/Entiteis/Cell.cs
@@ -6,7 +6,7 @@ namespace ProjectEntra21.src.Domain.Entiteis
 {
     public class Cell : PatternEntity
     {
-        public int CodeCell { get; set; }
+        public long CodeCell { get; set; }
         public StatusCell StatusCell { get; set; }
     }
 }
